+++ core/runner/src/main/java/com/cognifide/aet/runner/processing/data/wrappers/RunIndexWrapperFactory.java
@@ -13,7 +13,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.runner.processing.data.RunIndexWrappers;
+package com.cognifide.aet.runner.processing.data.wrappers;
 
 import com.cognifide.aet.communication.api.metadata.RunType;
 import com.cognifide.aet.communication.api.wrappers.Run;
