+++ core/runner/src/main/java/com/cognifide/aet/runner/processing/steps/ComparisonResultsRouter.java
@@ -28,7 +28,7 @@ import com.cognifide.aet.communication.api.util.ExecutionTimer;
 import com.cognifide.aet.runner.MessagesManager;
 import com.cognifide.aet.runner.RunnerConfiguration;
 import com.cognifide.aet.runner.processing.TimeoutWatch;
-import com.cognifide.aet.runner.processing.data.RunIndexWrappers.RunIndexWrapper;
+import com.cognifide.aet.runner.processing.data.wrappers.RunIndexWrapper;
 import java.util.Optional;
 import javax.jms.JMSException;
 import javax.jms.Message;
